Updated go test script and refactored collection package.
**Added:**

- `run-go-tests.sh`: Added check and cleanup for `go.mod` and `go.sum` before running tests.
- `Cmd` struct in `sysutils.go`: Added a mutex for thread-safe output handling.

**Changed:**

- Renamed package `slices` to `collection` for better clarity.
- Updated import paths in `slices_test.go` to reflect the new package name.
- Refactored `sysutils_test.go` to use pointer for `Cmd` struct.
- Cleaned up dependencies in `go.mod` and `go.sum` by removing unused ones.

**Removed:**

- Unused dependencies from `go.mod` and `go.sum` files.
